Add http:request directive
The http:request directive complements the http:response
directive.
It has the same fields as the http:response directive
(contenttype, data) and is meant to describe the body
of an http request.

This commit also changes the way http:response and
http:request directives are rendered. Their required
argument is now optional and the strings 'Reponse Body:'
or 'Request Body:' are now added in front of their
optional description.